cambiado tipo de field del argumento cedula  de la clase cliente

+++ backend/models.py
@@ -11,7 +11,7 @@ class usuario (models.Model):
 
 class cliente (models.Model):
 
-    cedula = models.IntegerField(validators=[MaxValueValidator(10000000000000)])
+    cedula = models.BigIntegerField(validators=[MaxValueValidator(10000000000000)])
 
 class sede (models.Model):
 
@@ -20,7 +20,7 @@ class sede (models.Model):
 
 class caja (models.Model):
 
-    numero_caja = models.BigIntegerField(validators=[MaxValueValidator(100)])
+    numero_caja = models.IntegerField(validators=[MaxValueValidator(100)])
     tipo = models.CharField(max_length=20)
 
 class sede_caja (models.Model):
